+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.9.3] - 2024-10-16
+
 ### Added
 
 - New `xuids` property on `MatchStats` as a convenience to get all user IDs.
@@ -147,7 +149,8 @@
 
 First documented release.
 
-[unreleased]: https://github.com/acurtis166/SPNKr/compare/v0.9.2...HEAD
+[unreleased]: https://github.com/acurtis166/SPNKr/compare/v0.9.3...HEAD
+[0.9.3]: https://github.com/acurtis166/SPNKr/compare/v0.9.2...v0.9.3
 [0.9.2]: https://github.com/acurtis166/SPNKr/compare/v0.9.1...v0.9.2
 [0.9.1]: https://github.com/acurtis166/SPNKr/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/acurtis166/SPNKr/compare/v0.8.0...v0.9.0